requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.1